backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
duckdb==1.1.3
pandas==2.2.3
sqlparse==0.5.1

# LangChain 0.3 line (stable)
langchain==0.3.8
langchain-community==0.3.8
langchain-experimental==0.3.4   # ← use an available version
langchain-google-genai==2.0.2
google-generativeai==0.8.3

# SQL / connectivity
sqlalchemy>=2.0
duckdb-engine>=0.11.0

# HTTP client
httpx>=0.27
rapidfuzz>=3.14.3